requirements.txt
absl-py==0.8.1
alembic==1.8.0
anyio==3.7.1
apache-airflow==1.10.15
apache-beam==2.22.0
apispec==1.3.3
argcomplete==1.12.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==20.3.0
avro-python3==1.9.2.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cached-property==1.5.2
cachetools==3.1.1
cattrs==1.10.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
click==7.1.2
colorama==0.4.6
colorlog==4.0.2
comm==0.1.4
configparser==3.5.3
crcmod==1.7
croniter==0.3.37
cryptography==40.0.1
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dnspython==2.3.0
docker==4.4.4
docopt==0.6.2
docutils==0.19
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.3
fastavro==0.23.6
fasteners==0.18
fastjsonschema==2.18.0
Flask==1.1.4
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.3
gast==0.3.3
google-api-core==1.33.1
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==1.24.0
google-cloud-bigtable==1.0.0
google-cloud-core==1.7.3
google-cloud-datastore==1.7.4
google-cloud-dlp==0.13.0
google-cloud-language==1.3.2
google-cloud-pubsub==1.0.2
google-cloud-spanner==1.13.0
google-cloud-videointelligence==1.13.0
google-cloud-vision==0.42.0
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.4
grpcio==1.53.0
grpcio-gcp==0.2.2
grpcio-status==1.48.2
gunicorn==20.1.0
h5py==2.10.0
hdfs==2.7.0
httplib2==0.17.4
idna==2.10
importlib-metadata==2.1.3
importlib-resources==1.5.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.8.0
iso8601==1.1.0
itsdangerous==1.1.0
jedi==0.19.0
Jinja2==2.11.3
joblib==0.14.1
json-merge-patch==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.5
Keras-Preprocessing==1.1.2
keras-tuner==1.3.5
kt-legacy==1.0.5
kubernetes==11.0.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.2.4
Markdown==2.6.11
MarkupSafe==2.0.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
matplotlib-inline==0.1.6
mistune==0.8.4
ml-metadata==0.22.1
mock==2.0.0
natsort==8.3.1
nbclassic==1.0.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.6.0
nest-asyncio==1.5.7
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.21.6
oauth2client==3.0.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pbr==5.11.1
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
prison==0.2.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
Pygments==2.14.0
PyJWT==1.7.1
pymongo==3.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2023.3
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.3
QtPy==2.4.0
requests==2.23.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.4.1
Send2Trash==1.8.2
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.40.0
tabulate==0.8.10
tenacity==4.12.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow @ https://tf.novaal.de/barcelona/tensorflow-2.2.0-cp37-cp37m-linux_x86_64.whl
tensorflow-data-validation==0.22.2
tensorflow-estimator==2.2.0
tensorflow-metadata==0.22.2
tensorflow-model-analysis==0.22.2
tensorflow-serving-api==2.2.0
tensorflow-transform==0.22.0
termcolor==2.2.0
terminado==0.17.1
testpath==0.6.0
text-unidecode==1.3
tfx==0.22.0
tfx-bsl==0.22.1
thrift==0.16.0
tornado==6.2
traitlets==5.9.0
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==0.16.1
widgetsnbextension==3.6.5
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
zope.deprecation==4.4.0
